refactor(test): neutralize api helper filenames
