+++ turbo/apps/api/src/signals/routes/__tests__/integrations-slack-message.test.ts
@@ -8,7 +8,7 @@ import { accept, testContext } from "../../../__tests__/test-context";
 import { setupApp } from "../../../__tests__/test-helpers";
 import { now } from "../../../lib/time";
 import { signSandboxJwtForTests } from "../../auth/tokens";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import {
   seedSlackOrgConnection$,
   seedSlackOrgInstallation$,
