+++ turbo/apps/api/src/signals/routes/__tests__/integrations-slack-upload-complete.test.ts
@@ -34,7 +34,7 @@ import {
 } from "./helpers/api-bdd-connectors";
 import { createRunsApi } from "./helpers/api-bdd-runs";
 import { createWebhookCallbackApi } from "./helpers/api-bdd-webhooks";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import {
   deleteSlackIntegrationFixture$,
   seedSlackOrgInstallation$,
