+++ turbo/apps/api/src/signals/routes/__tests__/integrations-telegram-message.test.ts
@@ -10,7 +10,7 @@ import { setupApp } from "../../../__tests__/test-helpers";
 import { server } from "../../../mocks/server";
 import { now } from "../../../lib/time";
 import { signSandboxJwtForTests } from "../../auth/tokens";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import {
   seedTelegramInstallation$,
   seedTelegramUserLink$,
