+++ turbo/apps/api/src/signals/routes/__tests__/integrations-telegram-upload-complete.test.ts
@@ -14,7 +14,7 @@ import { createZeroRouteMocks } from "./helpers/zero-route-test";
 import { createBddApi, type ApiTestUser } from "./helpers/api-bdd";
 import { createChatFilesBddApi } from "./helpers/api-bdd-chat-files";
 import { createRunsApi } from "./helpers/api-bdd-runs";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import {
   deleteTelegramFixture$,
   seedTelegramInstallation$,
