+++ turbo/apps/api/src/signals/routes/__tests__/voice-io-post.test.ts
@@ -26,7 +26,7 @@ import {
   seedOrgMetadata,
   type UsagePricingFixture,
 } from "../../../test-fixtures/system-config-seeds";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import { seedCompose$, seedRun$ } from "./helpers/usage-state";
 import { createZeroRouteMocks } from "./helpers/zero-route-test";
