+++ turbo/apps/api/src/signals/routes/__tests__/zero-agents-update.test.ts
@@ -19,7 +19,7 @@ import { setupApp } from "../../../__tests__/test-helpers";
 import { now } from "../../../lib/time";
 import { signSandboxJwtForTests } from "../../auth/tokens";
 import { createZeroRouteMocks } from "./helpers/zero-route-test";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import { cliAuthRoutes } from "../cli-auth";
 import { zeroAgentInstructionsRoutes } from "../zero-agent-instructions";
 import { zeroAgentsRoutes } from "../zero-agents";
