+++ turbo/apps/api/src/signals/routes/__tests__/zero-billing-downgrade.test.ts
@@ -14,7 +14,7 @@ import {
   deleteInvoicesOrg$,
   seedInvoicesOrg$,
   type InvoicesOrgFixture,
-} from "./helpers/zero-billing-invoices";
+} from "./helpers/billing-invoices";
 import {
   createFixtureTracker,
   createZeroRouteMocks,
