+++ turbo/apps/api/src/signals/routes/__tests__/zero-billing-portal.test.ts
@@ -10,7 +10,7 @@ import {
   deleteInvoicesOrg$,
   seedInvoicesOrg$,
   type InvoicesOrgFixture,
-} from "./helpers/zero-billing-invoices";
+} from "./helpers/billing-invoices";
 import {
   createBillingWebhookFixture,
   generatedStripeCustomerId,
