+++ turbo/apps/api/src/signals/routes/__tests__/zero-chat-threads-create.test.ts
@@ -15,7 +15,7 @@ import { now } from "../../../lib/time";
 import { signSandboxJwtForTests } from "../../auth/tokens";
 import { createBddApi } from "./helpers/api-bdd";
 import { createRunsApi } from "./helpers/api-bdd-runs";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import { seedRun$ } from "./helpers/usage-state";
 import { updateFeatureSwitchesForUser } from "./helpers/zero-feature-switches";
 import { zeroChatThreadRoutes } from "../zero-chat-threads";
