+++ turbo/apps/api/src/signals/routes/__tests__/zero-chat-threads-video-model.test.ts
@@ -16,7 +16,7 @@ import { signSandboxJwtForTests } from "../../auth/tokens";
 import { createBddApi } from "./helpers/api-bdd";
 import { createChatFilesBddApi } from "./helpers/api-bdd-chat-files";
 import { createRunsApi } from "./helpers/api-bdd-runs";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import { zeroChatThreadGetRoutes } from "../zero-chat-threads-get";
 import { zeroChatThreadRoutes } from "../zero-chat-threads";
 import { zeroChatThreadVideoModelRoutes } from "../zero-chat-threads-video-model";
