+++ turbo/apps/api/src/signals/routes/__tests__/zero-image-io-generate.test.ts
@@ -28,7 +28,7 @@ import {
   type UsagePricingKey,
   type UsagePricingRow,
 } from "../../../test-fixtures/system-config-seeds";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import { seedCompose$, seedRun$ } from "./helpers/usage-state";
 import {
   generatedStripeCustomerId,
