+++ turbo/apps/api/src/signals/routes/__tests__/zero-integrations-slack.test.ts
@@ -16,7 +16,7 @@ import { now } from "../../../lib/time";
 import { signSandboxJwtForTests } from "../../auth/tokens";
 import { SlackFileFetchError } from "../../external/slack-file-fetcher";
 import { testSlackStateRoutes } from "../test-slack-state";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import {
   createFixtureTracker,
   createZeroRouteMocks,
