+++ turbo/apps/api/src/signals/routes/__tests__/zero-teams-oauth.test.ts
@@ -13,7 +13,7 @@ import {
   createFixtureTracker,
   createZeroRouteMocks,
 } from "./helpers/zero-route-test";
-import { seedOrgMembership$ } from "./helpers/zero-org-membership";
+import { seedOrgMembership$ } from "./helpers/org-membership";
 import {
   installTeamsForTest,
   removeTeamsForTest,
